Added autoloader capabilities and classes.
Add: yasumi package. However it is not reflected in this commit, because for some reason I cannot add it.
Add: Faker package. It is used by yasumi. Also is not reflected in this commit, beause I cannot add it. Both of those packages should be cloned from respective repositories anyway.